css. calendar
added links to stylesheet for each html and created a separate stylesheet for the header since it was a lot of duplicated code and didn't make sense to write it out so many times. Also added the button for "downloading the calendar". It doesn't download currently still working on that